[warning cleanup] remove warning and useless files
Remove warning shows with -Wextra, and remove some really old
files...

+++ libpvkernel/src/include/pvkernel/rush/uchardetect/JpCntx.h
@@ -69,7 +69,7 @@ public:
     {
       mTotalRel++;
       //count this sequence to its category counter
-      mRelSample[jp2CharContext[mLastCharOrder][order]]++;
+      mRelSample[(int)jp2CharContext[(int)mLastCharOrder][order]]++;
     }
     mLastCharOrder = order;
   }
